src/scientific_domains.py
import jax
import jax.numpy as jnp
import flax.linen as nn
from typing import List, Optional
import pennylane as qml
import qutip as qt
import pyquil
from pyquil import Program, get_qc
from pyquil.gates import RX, RY, MEASURE
from qiskit import QuantumCircuit, execute, Aer
import biopython as bp
from Bio import SeqIO, Align
import neuropy  # Placeholder for a hypothetical neuroscience library

class ScientificDomains(nn.Module):
    features: List[int]
    activation: callable = nn.relu
    use_quantum: bool = False
    use_bioinformatics: bool = False
    use_neuroscience: bool = False
    quantum_library: str = 'pennylane'  # Options: 'pennylane', 'qutip', 'pyquil'

    @nn.compact
    def __call__(self, x):
        for feat in self.features[:-1]:
            x = nn.Dense(feat)(x)
            x = self.activation(x)

        if self.use_quantum:
            x = self.quantum_circuit(x)

        if self.use_bioinformatics:
            x = self.bioinformatics_analysis(x)

        if self.use_neuroscience:
            x = self.neuroscience_processing(x)

        return nn.Dense(self.features[-1])(x)

    def quantum_circuit(self, x):
        if self.quantum_library == 'pennylane':
            return self._pennylane_circuit(x)
        elif self.quantum_library == 'qutip':
            return self._qutip_circuit(x)
        elif self.quantum_library == 'pyquil':
            return self._pyquil_circuit(x)
        else:
            raise ValueError(f"Unsupported quantum library: {self.quantum_library}")

    def _pennylane_circuit(self, x):
        dev = qml.device("default.qubit", wires=1)
        @qml.qnode(dev)
        def circuit(params, x):
            qml.RX(x[0], wires=0)
            qml.RY(params[0], wires=0)
            return qml.expval(qml.PauliZ(0))

        params = jnp.array([0.1])
        return jnp.array([circuit(params, xi) for xi in x])

    def _qutip_circuit(self, x):
        def circuit(params, xi):
            psi = qt.basis(2, 0)  # Initial state |0>
            psi = qt.rx(xi[0]) * psi  # RX rotation
            psi = qt.ry(params[0]) * psi  # RY rotation
            return qt.expect(qt.sigmaz(), psi)  # Measure in Z basis

        params = jnp.array([0.1])
        return jnp.array([circuit(params, xi) for xi in x])

    def _pyquil_circuit(self, x):
        def circuit(params, xi):
            prog = Program()
            ro = prog.declare('ro', 'BIT', 1)
            prog.inst(RX(xi[0], 0))
            prog.inst(RY(params[0], 0))
            prog.inst(MEASURE(0, ro[0]))

            qc = get_qc('1q-qvm')
            results = qc.run_and_measure(prog, trials=100)
            return 1 - 2 * jnp.mean(results[0])  # Convert to expectation value

        params = jnp.array([0.1])
        return jnp.array([circuit(params, xi) for xi in x])

    def bioinformatics_analysis(self, x):
        # Basic sequence alignment using Biopython
        aligner = Align.PairwiseAligner()
        seq1 = SeqIO.read("sequence1.fasta", "fasta")
        seq2 = SeqIO.read("sequence2.fasta", "fasta")
        alignments = aligner.align(seq1.seq, seq2.seq)
        return jnp.array([float(alignment.score) for alignment in alignments])

    def neuroscience_processing(self, x):
        # Placeholder for neuroscience-specific processing
        # This could involve signal processing, brain imaging analysis, etc.
        return neuropy.process_signal(x)  # Hypothetical function

# Additional utility functions for scientific domains can be added here
