requirements.txt
jax>=0.4.10,<0.5.0
jaxlib>=0.4.10,<0.5.0
ml_dtypes>=0.2.0,<0.4.0
flax>=0.7.2,<0.8.0
optax>=0.1.7,<0.2.0
tensorflow-cpu>=2.16.1,<2.17.0
keras>=3.5.0,<3.6.0
gym>=0.26.2,<0.27.0
pytest>=7.4.0,<7.5.0
flake8>=6.0.0,<6.1.0
numpy>=1.23.5,<1.26.4
scipy>=1.10.1,<1.11.0
matplotlib>=3.7.1,<3.8.0
aif360>=0.4.0,<0.6.0
packaging>=23.1,<24.0
gast>=0.6.0,<0.7.0
wrapt>=1.16.0,<1.17.0
pennylane>=0.32.0,<0.33.0
ibm-watson-machine-learning>=1.0.257,<1.1.0
scikit-learn>=1.0,<1.2.0
pandas>=1.5.0,<1.6.0
adversarial-robustness-toolbox>=1.15.0,<1.19.0
lale>=0.7.0,<0.8.0
qutip>=4.7.1,<4.8.0
pyquil>=3.5.4,<3.6.0
qiskit>=0.43.0,<0.44.0
biopython>=1.81,<1.82
scikit-bio>=0.5.8,<0.6.0
ete3>=3.1.2,<3.2.0
xarray>=2023.5.0,<2023.6.0
torch>=2.0.1,<2.1.0
git+https://github.com/google-deepmind/alphafold.git@main#egg=alphafold
transformers>=4.18.0,<4.19.0
